+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - Use OpenTelemetry span name as fallback for transaction name ([#3557](https://github.com/getsentry/sentry-java/pull/3557))
   - In certain cases we were sending transactions as "<unlabeled transaction>" when using OpenTelemetry 
+- Add OpenTelemetry span data to Sentry span ([#3593](https://github.com/getsentry/sentry-java/pull/3593))
 
 ## 8.0.0-alpha.4